project/requirements.txt
Django==1.8.12
django-extensions==1.6.1
django-crispy-forms==1.6.0
django-registration-redux==1.4
Pillow==3.2.0
requests==2.9.1
gunicorn==19.4.5
psycopg2==2.6.1
django-storages==1.4.1
djangorestframework==3.3.3
